Cross-correlated number and brightness time series image classes.

The cc_number_ts_img and cc_brightness_ts_img classes are designed to hold objects which are images calculated from the cross-correlated number and brightness technique.

cc_number_ts_img(img, frames_per_set, overlapped, thresh, swaps, filt)

cc_brightness_ts_img(img, frames_per_set, overlapped, thresh, swaps, filt)

Arguments

img

The calculated cross-correlated number or brightness time series image series.

frames_per_set

The number of frames used in the calculation of each point in the cross-correlated number or brightness time series.

overlapped

A boolean. TRUE indicates that the windows used to calculate consecutive brightnesses over time were overlapped, FALSE indicates that they were not.

thresh

A positive integer, possibly an object of class autothresholdr::th. If the different channels of the image had different thresholds, this argument may be specified as a vector or list (of positive integers, possibly objects of class autothresholdr::th), one element for each channel.

swaps

A non-negative integer with an attribute auto. If the different channels of the image had different swaps, this argument may be specified as a list (of non-negative integers with attributes auto), one element for each channel. For undetrended images, set swaps = NA.

filt

A string, the filtering method used. Must be either "mean" or "median", or NA for no filtering. If the different channels of the image had different filters, this may be specified as a character vector, one element for each channel.

Value

An object of class cc_number_ts_img or cc_brightness_ts_img.

Details

An object of class cc_number_ts_img or cc_brightness_ts_img is a 4-dimensional array of real numbers in the mould of an ijtiff_img with 3 attributes:

swaps

A non-negative integer indicating the parameter used for Robin Hood detrending with an attribute auto which is a logical indicating whether or not the parameter was chosen automatically (in the multi-channel case, one swaps per channel is given).

frames_per_set

A positive integer detailing how many frames were used in the calculation of each point in the number or brightness time series.

overlapped

A boolean. TRUE indicates that the windows used to calculate consecutive brightnesses over time were overlapped, FALSE indicates that they were not.

See also