Number and brightness image classes.

The number_img and brightness_img classes are designed to hold objects which are images calculated from the number and brightness technique.

number_img(img, def, thresh, swaps, filt)

brightness_img(img, def, thresh, swaps, filt)

Arguments

img
The calculated number or brightness image.

def
The number or brightness definition used.

thresh
A positive integer, possibly an object of class autothresholdr::th. If the different channels of the image had different thresholds, this argument may be specified as a vector or list (of positive integers, possibly objects of class autothresholdr::th), one element for each channel.

swaps
A non-negative integer with an attribute auto. If the different channels of the image had different swaps, this argument may be specified as a list (of non-negative integers with attributes auto), one element for each channel. For undetrended images, set swaps = NA.

filt
A string, the filtering method used. Must be either "mean" or "median", or NA for no filtering. If the different channels of the image had different filters, this may be specified as a character vector, one element for each channel.

Value

An object of class number_img or brightness_img.

Details

An object of class number_img or brightness_img is a 4-dimensional array of real numbers in the mould of an ijtiff_img (indexed as img[y, x, channel, frame]) with 4 attributes:

thresh

A positive integer, possibly an object of class autothresholdr::th detailing which threshold and thresholding method was used in preprocessing (in the multi-channel case, one threshold per channel is given).

swaps

A non-negative integer indicating the number of swaps Robin Hood detrending, with an attribute auto which is a logical indicating whether or not the parameter was chosen automatically (in the multi-channel case, one threshold per channel is given).

filt

Was mean or median filtering used in postprocessing?